+++ probe-rs/src/architecture/arm/component/mod.rs
@@ -2,7 +2,9 @@
 
 mod dwt;
 mod itm;
+mod swo;
 mod tpiu;
+mod trace_funnel;
 
 use super::memory::romtable::{CoresightComponent, PeripheralType, RomTableError};
 use crate::architecture::arm::core::armv6m::Demcr;
@@ -10,7 +12,9 @@ use crate::architecture::arm::{ArmProbeInterface, SwoConfig, SwoMode};
 use crate::{Core, CoreRegister, Error, MemoryInterface};
 pub use dwt::Dwt;
 pub use itm::Itm;
+pub use swo::Swo;
 pub use tpiu::Tpiu;
+pub use trace_funnel::TraceFunnel;
 
 /// An error when operating a core ROM table component occurred.
 #[derive(thiserror::Error, Debug)]
@@ -112,6 +116,31 @@ pub(crate) fn setup_swv(
     } else {
         // Clear EnFCont to only pass through raw ITM/DWT data.
         tpiu.set_formatter(0x100)?;
+    }
+
+    // Configure SWO - it may not be present in some architectures, as the TPIU may drive SWO.
+    if let Ok(component) = find_component(components, PeripheralType::Swo) {
+        let mut swo = Swo::new(interface, component);
+        swo.unlock()?;
+
+        swo.set_prescaler(prescaler)?;
+
+        match config.mode() {
+            SwoMode::Manchester => swo.set_pin_protocol(1)?,
+            SwoMode::Uart => swo.set_pin_protocol(2)?,
+        }
+    } else {
+        log::warn!("SWO component not found - assuming TPIU-only configuration");
+    }
+
+    // Enable all ports of any trace funnels found.
+    for trace_funnel in components
+        .iter()
+        .filter_map(|comp| comp.find_component(PeripheralType::TraceFunnel))
+    {
+        let mut funnel = TraceFunnel::new(interface, trace_funnel);
+        funnel.unlock()?;
+        funnel.enable_port(0xFF)?;
     }
 
     // Configure ITM